+++ CHANGES.md
@@ -1,6 +1,8 @@
 
 - Safe-string support. Thanks to Josh Allmann for the help.
 - Deprecate `Uuidm.create` in favor of `Uuidm.v`.
+- Deprecate `Uuidm.print` in favor of `Uuidm.pp_string`
+- Add `Uuidm.pp`.
 - Relicensed from BSD3 to ISC.
 - Build depend on topkg.
 - `uuidtrip` uses `Cmdliner` which becomes a depopt of the